Ronica Peraka
I practiced clustering and kNN, and it was good to get more practice on the material. I tried to use the chooseK function to choose a k number, but I ended up looking at the graphs and visually counted clumps that could work. I learned to look at a new dataset that I didn't know. One thing I looked up is how to do clustering and how to use the chooseK function for data with non-numeric inputs. We have had experience with numbers and normalizing, but this data has characters and true/false.